fix: update tests to reflect change in handling jms reader
The existing tests were benefitting from the previous handling of
the JMS reader. Now that the reader is being closed (without the
reader committing what it's consumed from MQ) the messages
fetched from the queue manager are being rolled back at the end of
every test case when the source task is stopped. These messages
are then retrieved from MQ again in the subsequent test case which
is causing test failures.

This is a byproduct of the source task design - that only makes
commits to the MQ queue manager for messages retrieved in a poll()
call in a subsequent poll.

For normal operation this is okay, but for tests, we need to force
there to be a subsequent poll. This commit adds a consistent way
of doing this.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>